+++ src/data/index.js
@@ -1,2 +1,3 @@
 
+export * from "./getMockData";
 export * from "./OstrichImageData";