+++ src/tools/index.js
@@ -1,4 +1,5 @@
 
+export * from "./CursorInfoTool";
 export * from "./SliceScrollingTool";
 export * from "./Tool";
 export * from "./VoiWindowingTool";